Move context file to a separated folder
