+++ buildbox-app/src/components/feed.tsx
@@ -1,6 +1,6 @@
 import Post from './post'
 import styles from '../styles/feed.module.css'
-import { usePosts } from './PostsContext'
+import { usePosts } from '../contexts/PostsContext'
 
 const Feed = () => {
   const posts = usePosts()
