+++ buildbox-app/src/components/form.tsx
@@ -3,7 +3,7 @@ import Image from 'next/image'
 import uploadImage from '../../public/upload-image.png'
 import trashImage from '../../public/trash.png'
 import styles from '../styles/form.module.css'
-import { useDispatchPosts } from '../components/PostsContext'
+import { useDispatchPosts } from '../contexts/PostsContext'
 
 const Form = () => {
   const [photo, setPhoto] = useState('')
