+++ buildbox-app/src/pages/_app.tsx
@@ -1,6 +1,6 @@
 import '../styles/globals.css'
 import type { AppProps } from 'next/app'
-import { PostsProvider } from '../components/PostsContext'
+import { PostsProvider } from '../contexts/PostsContext'
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
